+++ sha256_mb/sha256_mb_test.c
@@ -184,11 +184,11 @@ main(void)
         }
 
         for (i = 0; i < MSGS; i++) {
-                ctx = sha256_ctx_mgr_submit(mgr, &ctxpool[i], msgs[i], strlen((char *) msgs[i]),
-                                            HASH_ENTIRE);
+                ctx = sha256_ctx_mgr_submit(mgr, &ctxpool[i], msgs[i],
+                                            (uint32_t) strlen((char *) msgs[i]), HASH_ENTIRE);
 
                 if (ctx) {
-                        t = (uintptr_t) (ctx->user_data);
+                        t = (uint32_t) (uintptr_t) (ctx->user_data);
                         good = expResultDigest[t];
                         checked++;
                         for (j = 0; j < SHA256_DIGEST_NWORDS; j++) {
@@ -212,7 +212,7 @@ main(void)
                 ctx = sha256_ctx_mgr_flush(mgr);
 
                 if (ctx) {
-                        t = (uintptr_t) (ctx->user_data);
+                        t = (uint32_t) (uintptr_t) (ctx->user_data);
                         good = expResultDigest[t];
                         checked++;
                         for (j = 0; j < SHA256_DIGEST_NWORDS; j++) {
@@ -245,10 +245,10 @@ main(void)
         checked = 0;
         for (i = 0; i < NUM_JOBS; i++) {
                 j = PSEUDO_RANDOM_NUM(i);
-                ctx = sha256_ctx_mgr_submit(mgr, &ctxpool[i], msgs[j], strlen((char *) msgs[j]),
-                                            HASH_ENTIRE);
+                ctx = sha256_ctx_mgr_submit(mgr, &ctxpool[i], msgs[j],
+                                            (uint32_t) strlen((char *) msgs[j]), HASH_ENTIRE);
                 if (ctx) {
-                        t = (uintptr_t) (ctx->user_data);
+                        t = (uint32_t) (uintptr_t) (ctx->user_data);
                         k = PSEUDO_RANDOM_NUM(t);
                         good = expResultDigest[k];
                         checked++;
@@ -267,7 +267,7 @@ main(void)
                                 goto end;
                         }
 
-                        t = (uintptr_t) (ctx->user_data);
+                        t = (uint32_t) (uintptr_t) (ctx->user_data);
                         k = PSEUDO_RANDOM_NUM(t);
                 }
         }
@@ -275,7 +275,7 @@ main(void)
                 ctx = sha256_ctx_mgr_flush(mgr);
 
                 if (ctx) {
-                        t = (uintptr_t) (ctx->user_data);
+                        t = (uint32_t) (uintptr_t) (ctx->user_data);
                         k = PSEUDO_RANDOM_NUM(t);
                         good = expResultDigest[k];
                         checked++;
