+++ sha512_mb/sha512_mb_rand_update_test.c
@@ -73,7 +73,7 @@ main(void)
         SHA512_HASH_CTX_MGR *mgr = NULL;
         SHA512_HASH_CTX ctxpool[TEST_BUFS], *ctx = NULL;
         uint32_t i, j, fail = 0;
-        int len_done, len_rem, len_rand;
+        uint32_t len_done, len_rem, len_rand;
         unsigned char *bufs[TEST_BUFS];
         unsigned char *buf_ptr[TEST_BUFS];
         uint32_t lens[TEST_BUFS];
@@ -114,7 +114,7 @@ main(void)
 
         // Run sb_sha512 tests
         for (i = 0; i < TEST_BUFS;) {
-                len_done = (int) ((uintptr_t) buf_ptr[i] - (uintptr_t) bufs[i]);
+                len_done = (uint32_t) ((uintptr_t) buf_ptr[i] - (uintptr_t) bufs[i]);
                 len_rem = TEST_LEN - len_done;
 
                 if (len_done == 0)
@@ -227,7 +227,7 @@ main(void)
                                         buf_ptr[j] = bufs[j] + ctx->total_length;
                                         len_rand = (rand() % SHA512_BLOCK_SIZE) *
                                                    (rand() % MAX_RAND_UPDATE_BLOCKS);
-                                        len_rem = lens[j] - ctx->total_length;
+                                        len_rem = lens[j] - (uint32_t) ctx->total_length;
 
                                         if (len_rem <=
                                             len_rand) // submit the rest of the job as LAST
@@ -257,7 +257,7 @@ main(void)
                         // Resubmit unfinished job
                         i = (unsigned long) (uintptr_t) (ctx->user_data);
                         buf_ptr[i] = bufs[i] + ctx->total_length; // update buffer pointer
-                        len_rem = lens[i] - ctx->total_length;
+                        len_rem = lens[i] - (uint32_t) ctx->total_length;
                         len_rand = (rand() % SHA512_BLOCK_SIZE) * (rand() % MAX_RAND_UPDATE_BLOCKS);
                         debug_char('+');
                         if (len_rem <= len_rand)
